python38-313: fix threadid patch
The patch addresses a bug in Python code which upstream does not bother
fixing: https://github.com/python/cpython/pull/113890
MacPorts patch has an error which upstream does not want to fix either:
https://github.com/macports/macports-ports/pull/20761
So we are stuck with carrying this indefinitely.
Notice, this is not specific to 10a190, but also needed on 10.6.8,
at least until someone fixes its libSystem.