_resouces: include all portgroups which are used by MacPorts
These are non-modified PGs from upstream MacPorts, which are included
in order for the first port sync run to be successful.
The issue is that powerpc-ports are synced first, and they need
mainstream PGs to be present. This is not an issue on subsequent runs,
but it is undesirable for the initial run to result in multiple errors.