MacPorts: allow clangs on 10.7+, fix gcc choice after recent upstream breakage
This allows PPCPorts to be used normally on non-powerpc systems, at least macOS.
It also avoids pulling in gcc15 via gcc_compilers file, which happened after recent upstream commit. There is nothing wrong with gcc15, but there is no reason to hurry with switching to it, and compiler choice must be coherent.