mldonkey: fix build on Tiger
The 'const char **' error is common on Tiger.  An example with the solution I borrowed can be found here: https://trac.macports.org/ticket/71658
I can present the failed log if you think that might lead to a better solution, but mldonkey is working fine for me. Web GUI launches and is navigable. I don't know enough about these types of programs to know if all functionality is there.